Obsolete ConstructionTests by ensuring all scenarios are covered by LifecycleTests.
